requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
gcloud==0.17.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
jws==0.1.3
oauth2client==3.0.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.9
